+++ tests/Pest.php
@@ -3,3 +3,4 @@
 use MessengerBot\Tests\TestCase;
 
 pest()->extend(TestCase::class)->in(__DIR__.'/Feature');
+pest()->extend(TestCase::class)->in(__DIR__.'/Unit');